Fix i32x4 left shift by integer (i32) function
The shift amount mustn't be copied in each of the 4-byte integers of the
m128i value. Instead the whole m128i value represents one single integer,
and each 4-byte integer is shifted the same amount.

Also added an indexing function to i32x4, which is used in the tests.