Change target folder for native cpu build in CI
GitHub Actions seems to have started scanning executables. However while
the scan is still in progress, we redo the build for the native CPU.
Cargo will then try to replace the previously built executable and fail
because the scan is still in progress. By switching the target folder
for the native build, we can avoid this problem.