Improve performance of saturating float to int casts (#108)
* Change target folder for native cpu build in CI

GitHub Actions seems to have started scanning executables. However while
the scan is still in progress, we redo the build for the native CPU.
Cargo will then try to replace the previously built executable and fail
because the scan is still in progress. By switching the target folder
for the native build, we can avoid this problem.

* Improve performance of saturating float to int casts

I randomly stumbled across a Chrome bug ticket where they reported that
their lowering of saturating float to int casts is suboptimal:

https://bugs.chromium.org/p/v8/issues/detail?id=12094

They by now merged a fix. The code I ported was based on Chrome and thus
the same performance pitfall applies to wide.