fix serialize impl of SIMD types
change serialization of SIMD types to use serde::Serializer::serialize_tuple

The Serialize implementation of all types used Serializer::serialize_seq
whereas the deserialize implementation delegated to that of the
corresponding [T; LEN] type, which is implemented by serde using
serialize_tuple. Formats like bincode will emit the lenght of the
sequence but not of a tuple, leading to wrong deserialization.